requirements.txt
Django==4.2
python-dotenv==1.0.1
psycopg2==2.9.9
djangorestframework==3.14.0
django-filter==24.2
django-phonenumber-field==7.2.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
phonenumbers==8.13.26
Pillow==10.2.0
psycopg2-binary==2.9.9
